raspberry_pi/requirements.txt
smbus-cffi==0.5.1
RPi.GPIO==0.7.0
paho-mqtt==1.5.0